+++ pastell-core/RechercheAvanceFormulaireHTML.php
@@ -191,8 +191,7 @@ class RechercheAvanceFormulaireHTML extends PastellControler
 
     private function displayEtatTransit()
     {
-        $allDroit = $this->getInstance(RoleUtilisateur::class)->getAllDroit($this->getId_u());
-        $listeEtat = $this->getInstance(DocumentTypeFactory::class)->getActionByRole($allDroit);
+        $listeEtat = $this->getDocumentStateList();
         $etatTransit = $this->getParameter('etatTransit');
         ?>
         <select name='etatTransit' class="form-control col-md-8">
@@ -222,8 +221,7 @@ class RechercheAvanceFormulaireHTML extends PastellControler
 
     private function displayNotEtatTransit()
     {
-        $allDroit = $this->getInstance(RoleUtilisateur::class)->getAllDroit($this->getId_u());
-        $listeEtat = $this->getInstance(DocumentTypeFactory::class)->getActionByRole($allDroit);
+        $listeEtat = $this->getDocumentStateList();
         $notEtatTransit = $this->getParameter('notEtatTransit');
         ?>
         <select name='notEtatTransit' class="form-control col-md-8">
@@ -255,8 +253,7 @@ class RechercheAvanceFormulaireHTML extends PastellControler
 
     private function displayLastState()
     {
-        $allDroit = $this->getInstance(RoleUtilisateur::class)->getAllDroit($this->getId_u());
-        $listeEtat = $this->getInstance(DocumentTypeFactory::class)->getActionByRole($allDroit);
+        $listeEtat = $this->getDocumentStateList();
         $lastEtat = $this->getParameter('lastetat');
         ?>
         <select name='lastetat' class="form-control col-md-8">
@@ -353,5 +350,15 @@ class RechercheAvanceFormulaireHTML extends PastellControler
             });
         </script>
         <?php
+    }
+
+    public function getDocumentStateList(): array
+    {
+        $allDroit = $this->getInstance(RoleUtilisateur::class)->getAllDroit($this->getId_u());
+        $documentStateList = $this->getInstance(DocumentTypeFactory::class)->getActionByRole($allDroit);
+        foreach ($documentStateList as $stateKey => $state) {
+            $documentStateList[$stateKey]['fatal-error'] = 'Erreur fatale';
+        }
+        return $documentStateList;
     }
 }